+++ Sandbox/Sandbox/Initialization/InitializationView.swift
@@ -73,7 +73,7 @@ struct InitializationView: View {
                                 Text("Advanced")
                             }
                         }
-                        
+
                         Section(header: Text("Debug Tools")) {
                             NavigationLink(destination: SdkInitializationView()) {
                                 Text("Manual SDK Initialization")