+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # develop
 
 ## New features
+- BDN-976 Refactor Ad Loading Logic Based on SDK Initialization State
 - BDN-978 Initialization failure state
 - BDN-942 New Bidon SDK init logic
 - BDN-975 GitHub Actions